Move the hexdump header into the nicely namespaced directory
