+++ tools/tests-compress.c
@@ -19,13 +19,13 @@
 
 
 #include <assert.h>    // for assert
+#include <n3n/hexdump.h>  // for fhexdump
 #include <n3n/logging.h> // for traceEvent
 #include <stdint.h>    // for uint8_t
 #include <stdio.h>     // for printf, fprintf, stderr, stdout, NULL
 #include <stdlib.h>    // for exit
 #include <string.h>    // for memcmp
 #include "config.h"    // for HAVE_LIBZSTD
-#include "hexdump.h"   // for fhexdump
 #include "minilzo.h"   // for lzo1x_1_compress, lzo1x_decompress, LZO1X_1_ME...
 #include "n2n.h"       // for N2N_PKT_BUF_SIZE
